requirements.txt
transformers
torch
duckduckgo-search 
python-dotenv
serpapi
requests